refactor: component names for better clarity
