+++ src/app/(home)/(profile)/_components/setting/SettingsList.tsx
@@ -5,10 +5,10 @@ import PasswordModal from './profilemodal/PasswordModal';
 import NicknameModal from './profilemodal/NicknameModal';
 import InfoModal from './profilemodal/InfoModal';
 import { useState } from 'react';
-import DetailItem from './DetailItem';
+import SettingItem from './SettingItem';
 import OAuthLoginStatus from '@/app/(home)/(auth)/_components/OAuthLoginStatus';
 
-type ProfileDetailsProps = {
+type SettingsListProps = {
   nickname: string;
   email?: string;
   info: string;
@@ -16,19 +16,19 @@ type ProfileDetailsProps = {
   onInfoUpdate: (newInfo: string) => void;
 };
 
-const ProfileDetails = ({ nickname, email, info, onNicknameUpdate, onInfoUpdate }: ProfileDetailsProps) => {
+const SettingsList = ({ nickname, email, info, onNicknameUpdate, onInfoUpdate }: SettingsListProps) => {
   const [isPasswordModalOpen, setPasswordModalOpen] = useState<boolean>(false);
   const [isNicknameModalOpen, setNicknameModalOpen] = useState<boolean>(false);
   const [isInfoModalOpen, setInfoModalOpen] = useState<boolean>(false);
 
   return (
     <div className="w-[588px]">
-      <DetailItem label="이메일" value={email} />
+      <SettingItem label="이메일" value={email} />
       <OAuthLoginStatus />
       <p className="border-b border-neutral-50 " />
-      <DetailItem label="비밀번호" value="변경하기" onClick={() => setPasswordModalOpen(true)} />
-      <DetailItem label="닉네임" value={nickname} onClick={() => setNicknameModalOpen(true)} />
-      <DetailItem label="자기소개" value={filterSlang(info)} onClick={() => setInfoModalOpen(true)} />
+      <SettingItem label="비밀번호" value="변경하기" onClick={() => setPasswordModalOpen(true)} />
+      <SettingItem label="닉네임" value={nickname} onClick={() => setNicknameModalOpen(true)} />
+      <SettingItem label="자기소개" value={filterSlang(info)} onClick={() => setInfoModalOpen(true)} />
       <PasswordModal isOpen={isPasswordModalOpen} onClose={() => setPasswordModalOpen(false)} />
       <NicknameModal
         isOpen={isNicknameModalOpen}
@@ -46,4 +46,4 @@ const ProfileDetails = ({ nickname, email, info, onNicknameUpdate, onInfoUpdate 
   );
 };
 
-export default ProfileDetails;
+export default SettingsList;
